I did this some time ago just for fun
I have not yet implemented the OpenGL part for visuals and I do not know when I am gonna do it since it was more of a struggle that I expected. Also there may be some mistakes in some of my implementations. The Neural Networks is very slow due to the fact that it is designed to support in the OpenGL environment.